Start Tuftemark special converter
It is supposed to be used to convert Markdown files into HTML that is
easy to be used with Tufte CSS set of styles.

To get there, we need to make some transformations to the parsed
Markdown AST and apply various changes to attributes or even whole
blocks.

We start with simple task – wrap the whole Markdown content with a
single article container around.