Add conversion of citation-blockquotes
For blockquotes that presents a citation, Tufte CSS expects to have last
child element as `footer` + `a` inside it.

We want to simplify the approach we shall use in the original Markdown
and decided to use the same Kramdown syntax for attributes: basically,
in a Markdown we need to add `{:cite="https://example.com"}` kind of
line right after the blockquote block.

After that our code will try to automatically pick up such a blockquote
and restructure it accordingly:

- if original Markdown contains a paragraph of text as blockquote's last
  element, content of it will be take and used in the `a` tag
- if original Markdown doesn't have a paragaph (e.g., blockquote's block
  ends with a list of items), but we still instructed it to be a
  citation, then our code uses href value itself for `a` tag content

See conversion examples in the corresponding test suite.