Add conversion of footnotes (into margin ones)
In Tufte CSS, we do not usually use FOOT-notes because we mostly rely on
the marginal ones. (More or less) standard Markdown's conversion – which
collects all footnotes, generates a div-container of them and puts it to
the end of document (and adding some internal links to them) – is not
what we need.

Basically, we need to replace a footnote in `...Bembo[^bembo] font...`
and the following paragraph (that provides actual footnote's content)
with this set of tags:

```
...Bembo<label for="sn-bembo" class="margin-toggle sidenote-number"></label>
          <input type="checkbox" id="sn-bembo" class="margin-toggle">
          <span class="sidenote">See Tufte’s comment in the <a
          href="http://www.edwardtufte.com/bboard/q-and-a-fetch-msg?msg_id=0000Vt">Tufte
          book fonts</a> thread.</span> font...
```

For this (a little bit not a trivial) conversion we have to parse AST
twice:

- first time to collect the footnotes paragraphs (they have a special
  starting sequence like `[^bembo]: lorem ipsum...`)
- then we go over the (cleaned) AST again, check strings that contain
  footnotes, and attempt to replace these tiny pieces with a
  Tufte-complaint marginal note layout

P.S. We also decided to offload this logic to a dedicated module, so
probably, we will do a similar thing to the other Tuftemark-conversions.