Solidify sidenotes processor: refactor
With the attempt to introduce conversion of images and writing a first
test for them, our sidenotes processor broke, because it expects that a
paragraph starts with a text (to parse its beginning and see if it's a
marginal note or not).

With images, some paragraphs can start with `<img .../>` tags though...

In addition, some paragraphs can start with other tags and this would
break the processor too.

We used `with/else` form in order to make some extra checks before
parsing the content of a given paragraph.