Support epigraphs (version of blockquotes)
Previously, we have implemented Tuftemark.Citations module that supposed
to convert blockquotes with a special `cite` attribute set (via Kramdown
syntax).

Other blockquotes were left as they are.

However, Tufte CSS has a special support for epigraphs, which are built
from blockquotes.

We decided that we will have to rely on the `role` attribute which can
be set in Markdown for a given blockquote to mark it as an epigraph. It
must consist of **at least** two paragraphs to be converted properly to
HTML because our processor converts the last `p` tag into `footer` one.

P.S. In addition to the new processor, we also refactored (and renamed)
old Citations module to a more generic Blockquotes.