Use `.epigraph` class instead of blockquote's role
...to convert a blockquote to an epigraph style. We do this to align
with the rest of the postprocessor's convention.

The only exception now is the `cite` attribute that we set to a
blockquote to make it a citation, though, it's quite logical (and we
must provide a link/reference to a citation any way, so we have to rely
on an attribute).