Fix extra wrapping of text child els in Sidenotes
The issue related to the unintented conversion of binary elements into a
binary wrapped in a list due to `String.split/2` processing. Initially,
we wanted to split the string and replace only "footnotes" substrings
with the set of elements we created for that. Though, this split also
lead to an extra wrapping of string elements which has no a "footenote".

P.S. In addition, we also removed `@tag :focus` from all the tests.