Refactor Tuftemark.Sidenotes processor
As an intermediary step for the coming changes to Blockquotes processor,
we found (and eventually fixed) an annoying issue in the Sidenotes one:

In some cases, we get nested lists of children elements (because we
processed text nodes previously and replaced them with lists of marginal
notes elements).

So, we fixed that and made Sidenotes module code cleaner by giving
better names to the arguments, variables, functions, splitting code into
more granular logical pieces, etc.

Enjoy! :)